third_party/blink/web_tests/virtual/pna-permission/external/wpt/fetch/private-network-access/mixed-content-fetch.tentative.https.window_include=from-private-expected.txt

This is a testharness.js-based test.
PASS private to local with local targetAddressSpace option: success.
FAIL private to local with local targetAddressSpace option: PUT success. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL private to local with private targetAddressSpace option: failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
FAIL private to local with private targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS private to local with public targetAddressSpace option: failed.
PASS private to local with public targetAddressSpace option: PUT failed.
PASS private to private with local targetAddressSpace option: failed.
FAIL private to private with local targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS private to private with private targetAddressSpace option: failed.
FAIL private to private with private targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS private to private with public targetAddressSpace option: failed.
PASS private to private with public targetAddressSpace option: PUT failed.
PASS private to public with local targetAddressSpace option: failed.
FAIL private to public with local targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS private to public with private targetAddressSpace option: failed.
FAIL private to public with private targetAddressSpace option: PUT failed. assert_equals: error expected (string) "TypeError: Failed to fetch" but got (undefined) undefined
PASS private to public with public targetAddressSpace option: failed.
PASS private to public with public targetAddressSpace option: PUT failed.
Harness: the test ran to completion.
